+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/approval_rule/ApprovalRuleSetCustomFieldActionBuilder.java
@@ -1,0 +1,109 @@
+
+package com.commercetools.api.models.approval_rule;
+
+import java.util.*;
+
+import javax.annotation.Nullable;
+
+import io.vrap.rmf.base.client.Builder;
+import io.vrap.rmf.base.client.utils.Generated;
+
+/**
+ * ApprovalRuleSetCustomFieldActionBuilder
+ * <hr>
+ * Example to create an instance using the builder pattern
+ * <div class=code-example>
+ * <pre><code class='java'>
+ *     ApprovalRuleSetCustomFieldAction approvalRuleSetCustomFieldAction = ApprovalRuleSetCustomFieldAction.builder()
+ *             .name("{name}")
+ *             .build()
+ * </code></pre>
+ * </div>
+ */
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class ApprovalRuleSetCustomFieldActionBuilder implements Builder<ApprovalRuleSetCustomFieldAction> {
+
+    private String name;
+
+    @Nullable
+    private java.lang.Object value;
+
+    /**
+     *  <p>Name of the Custom Field.</p>
+     * @param name value to be set
+     * @return Builder
+     */
+
+    public ApprovalRuleSetCustomFieldActionBuilder name(final String name) {
+        this.name = name;
+        return this;
+    }
+
+    /**
+     *  <p>If <code>value</code> is absent or <code>null</code>, this field will be removed if it exists. Removing a field that does not exist returns an InvalidOperation error. If <code>value</code> is provided, it is set for the field defined by <code>name</code>.</p>
+     * @param value value to be set
+     * @return Builder
+     */
+
+    public ApprovalRuleSetCustomFieldActionBuilder value(@Nullable final java.lang.Object value) {
+        this.value = value;
+        return this;
+    }
+
+    /**
+     *  <p>Name of the Custom Field.</p>
+     * @return name
+     */
+
+    public String getName() {
+        return this.name;
+    }
+
+    /**
+     *  <p>If <code>value</code> is absent or <code>null</code>, this field will be removed if it exists. Removing a field that does not exist returns an InvalidOperation error. If <code>value</code> is provided, it is set for the field defined by <code>name</code>.</p>
+     * @return value
+     */
+
+    @Nullable
+    public java.lang.Object getValue() {
+        return this.value;
+    }
+
+    /**
+     * builds ApprovalRuleSetCustomFieldAction with checking for non-null required values
+     * @return ApprovalRuleSetCustomFieldAction
+     */
+    public ApprovalRuleSetCustomFieldAction build() {
+        Objects.requireNonNull(name, ApprovalRuleSetCustomFieldAction.class + ": name is missing");
+        return new ApprovalRuleSetCustomFieldActionImpl(name, value);
+    }
+
+    /**
+     * builds ApprovalRuleSetCustomFieldAction without checking for non-null required values
+     * @return ApprovalRuleSetCustomFieldAction
+     */
+    public ApprovalRuleSetCustomFieldAction buildUnchecked() {
+        return new ApprovalRuleSetCustomFieldActionImpl(name, value);
+    }
+
+    /**
+     * factory method for an instance of ApprovalRuleSetCustomFieldActionBuilder
+     * @return builder
+     */
+    public static ApprovalRuleSetCustomFieldActionBuilder of() {
+        return new ApprovalRuleSetCustomFieldActionBuilder();
+    }
+
+    /**
+     * create builder for ApprovalRuleSetCustomFieldAction instance
+     * @param template instance with prefilled values for the builder
+     * @return builder
+     */
+    public static ApprovalRuleSetCustomFieldActionBuilder of(final ApprovalRuleSetCustomFieldAction template) {
+        ApprovalRuleSetCustomFieldActionBuilder builder = new ApprovalRuleSetCustomFieldActionBuilder();
+        builder.name = template.getName();
+        builder.value = template.getValue();
+        return builder;
+    }
+
+}