+++ commercetools/commercetools-sdk-java-api/src/main/java-predicates-generated/com/commercetools/api/predicates/query/approval_rule/ApprovalRuleQueryBuilderDsl.java
@@ -118,4 +118,13 @@ public class ApprovalRuleQueryBuilderDsl {
             ApprovalRuleQueryBuilderDsl::of);
     }
 
+    public CombinationQueryPredicate<ApprovalRuleQueryBuilderDsl> custom(
+            Function<com.commercetools.api.predicates.query.type.CustomFieldsQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.type.CustomFieldsQueryBuilderDsl>> fn) {
+        return new CombinationQueryPredicate<>(
+            ContainerQueryPredicate.of()
+                    .parent(ConstantQueryPredicate.of().constant("custom"))
+                    .inner(fn.apply(com.commercetools.api.predicates.query.type.CustomFieldsQueryBuilderDsl.of())),
+            ApprovalRuleQueryBuilderDsl::of);
+    }
+
 }