+++ commercetools/commercetools-sdk-java-api/src/main/java-predicates-generated/com/commercetools/api/predicates/query/approval_rule/ApprovalRuleSetCustomTypeActionQueryBuilderDsl.java
@@ -1,0 +1,41 @@
+
+package com.commercetools.api.predicates.query.approval_rule;
+
+import java.util.function.Function;
+
+import com.commercetools.api.predicates.query.*;
+
+public class ApprovalRuleSetCustomTypeActionQueryBuilderDsl {
+    public ApprovalRuleSetCustomTypeActionQueryBuilderDsl() {
+    }
+
+    public static ApprovalRuleSetCustomTypeActionQueryBuilderDsl of() {
+        return new ApprovalRuleSetCustomTypeActionQueryBuilderDsl();
+    }
+
+    public StringComparisonPredicateBuilder<ApprovalRuleSetCustomTypeActionQueryBuilderDsl> action() {
+        return new StringComparisonPredicateBuilder<>(
+            BinaryQueryPredicate.of().left(new ConstantQueryPredicate("action")),
+            p -> new CombinationQueryPredicate<>(p, ApprovalRuleSetCustomTypeActionQueryBuilderDsl::of));
+    }
+
+    public CombinationQueryPredicate<ApprovalRuleSetCustomTypeActionQueryBuilderDsl> type(
+            Function<com.commercetools.api.predicates.query.type.TypeResourceIdentifierQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.type.TypeResourceIdentifierQueryBuilderDsl>> fn) {
+        return new CombinationQueryPredicate<>(
+            ContainerQueryPredicate.of()
+                    .parent(ConstantQueryPredicate.of().constant("type"))
+                    .inner(fn.apply(
+                        com.commercetools.api.predicates.query.type.TypeResourceIdentifierQueryBuilderDsl.of())),
+            ApprovalRuleSetCustomTypeActionQueryBuilderDsl::of);
+    }
+
+    public CombinationQueryPredicate<ApprovalRuleSetCustomTypeActionQueryBuilderDsl> fields(
+            Function<com.commercetools.api.predicates.query.type.FieldContainerQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.type.FieldContainerQueryBuilderDsl>> fn) {
+        return new CombinationQueryPredicate<>(
+            ContainerQueryPredicate.of()
+                    .parent(ConstantQueryPredicate.of().constant("fields"))
+                    .inner(fn.apply(com.commercetools.api.predicates.query.type.FieldContainerQueryBuilderDsl.of())),
+            ApprovalRuleSetCustomTypeActionQueryBuilderDsl::of);
+    }
+
+}